Semantics-New/pointer02.c: test files was wrong and outdated, source code was buggy
git-svn-id: https://scm.cri.ensmp.fr/svn/validation/trunk@5948 67b5017d-e62d-0410-a871-f72b1c62810f

+++ Semantics-New/pointer02.c
@@ -17,7 +17,7 @@ void foo(double *p)
 int main()
 {
   x= 1, y=2, z=3;
-  double *p = (int *)malloc(sizeof(double));
+  double *p = (double *)malloc(sizeof(double));
   foo(p);
   return x+y+z; // information about x, y and z should be preserved
 }
